requirements.txt
Django==5.2.7
djangorestframework==3.15.2
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
django-redis==5.4.0
boto3==1.34.34
pandas==2.2.0
scikit-learn==1.4.0
openpyxl==3.1.2
xlrd==2.0.1
